322: Generate base URI for links in emails from HTTP_REFERER header

+++ backend/entrypoint.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 rm -rf /app/tmp/**/*
 
-bundle install --clean
+# Only re-install if Gemfile.lock is missing or empty
+if [[ ! -f Gemfile.lock  || ! -s Gemfile.lock ]]; then
+    bundle install --clean
+fi
 
 bundle exec rails server -p $PORT -b '0.0.0.0'
